create a savedstateflow-hilt module that automatically scopes SavedStateFlowHandle to the hilt ViewModelScope
