+++ .github/workflows/publish.yml
@@ -20,7 +20,7 @@ jobs:
 
         # Builds the release artifacts of the library
       - name: Release build
-        run: ./gradlew :savedstateflow:assembleRelease :savedstateflow-test:assembleRelease
+        run: ./gradlew :savedstateflow:assembleRelease :savedstateflow-hilt:assembleRelease :savedstateflow-test:assembleRelease
 
         # Generates other artifacts (javadocJar is optional)
       - name: Source jar and dokka
