+++ sample/hilt-di/build.gradle
@@ -48,7 +48,7 @@ android {
 }
 
 dependencies {
-    implementation(project(":savedstateflow"))
+    implementation(project(":savedstateflow-hilt"))
     implementation "androidx.core:core-ktx:$core_ktx_version"
     implementation "androidx.compose.ui:ui:$compose_version"
     implementation "androidx.compose.material:material:$compose_version"
