+++ sample/hilt-di/src/main/java/com/plusmobileapps/savedstateflow/Modules.kt
@@ -9,17 +9,6 @@ import dagger.hilt.android.components.ViewModelComponent
 import dagger.hilt.android.scopes.ViewModelScoped
 import dagger.hilt.components.SingletonComponent
 
-@InstallIn(ViewModelComponent::class)
-@Module
-object SavedStateFlowHandleModule {
-
-    @Provides
-    @ViewModelScoped
-    fun providesSavedStateFlowHandle(savedStateHandle: SavedStateHandle): SavedStateFlowHandle =
-        savedStateHandle.toSavedStateFlowHandle()
-
-}
-
 @Module
 @InstallIn(SingletonComponent::class)
 abstract class NewsModule {
