+++ savedstateflow-hilt/src/main/AndroidManifest.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.plusmobileapps.savedstateflow_hilt">
+
+</manifest>
