+++ savedstateflow-hilt/src/main/java/com/plusmobileapps/savedstateflow_hilt/SavedStateFlowHandleModule.kt
@@ -1,0 +1,21 @@
+package com.plusmobileapps.savedstateflow_hilt
+
+import androidx.lifecycle.SavedStateHandle
+import com.plusmobileapps.savedstateflow.SavedStateFlowHandle
+import com.plusmobileapps.savedstateflow.toSavedStateFlowHandle
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ViewModelComponent
+import dagger.hilt.android.scopes.ViewModelScoped
+
+@InstallIn(ViewModelComponent::class)
+@Module
+object SavedStateFlowHandleModule {
+
+    @Provides
+    @ViewModelScoped
+    fun providesSavedStateFlowHandle(savedStateHandle: SavedStateHandle): SavedStateFlowHandle =
+        savedStateHandle.toSavedStateFlowHandle()
+
+}
