+++ settings.gradle
@@ -17,3 +17,4 @@ include ':sample:manual-di'
 include ':sample:hilt-di'
 include ':savedstateflow'
 include ':savedstateflow-test'
+include ':savedstateflow-hilt'
